+++ lib/src/util/logger.dart
@@ -1,0 +1,4 @@
+import 'package:logging/logging.dart';
+
+final Logger databaseLogger = Logger('rinne_graph.database');
+final Logger traversalLogger = Logger('rinne_graph.traversal');